move types to mod.rs
