+++ crates/webauthn/tests/src/auth/mod.rs
@@ -2,9 +2,66 @@ use cairo_args_runner::{arg_array, arg_value, felt_vec, Arg, Felt252};
 
 mod mod_arithmetic;
 mod verify_signature;
+mod verify_ecdsa;
 
+#[derive(Debug)]
 pub struct ArgsBuilder {
     args: Vec<Arg>,
+}
+
+pub struct CairoU256 {
+    low: Felt252,
+    high: Felt252,
+}
+
+
+impl CairoU256 {
+    pub fn new(low: Felt252, high: Felt252) -> Self {
+        Self { low, high }
+    }
+    pub fn from_bytes_be(low: &[u8; 16], high: &[u8; 16]) -> Self {
+        Self::new(Felt252::from_bytes_be(low), Felt252::from_bytes_be(high))
+    }
+    pub fn from_byte_slice_be(bytes: &[u8; 32]) -> Self {
+        let (low, high): (&[u8; 16], &[u8; 16]) = (
+            bytes[16..].try_into().unwrap(),
+            bytes[..16].try_into().unwrap(),
+        );
+        Self::from_bytes_be(low, high)
+    }
+}
+
+impl FeltSerialize for CairoU256 {
+    fn to_felts(self) -> Vec<Felt252> {
+        vec![self.low, self.high]
+    }
+}
+
+struct P256r1PublicKey {
+    x: CairoU256,
+    y: CairoU256,
+}
+
+impl P256r1PublicKey {
+    pub fn new(x: CairoU256, y: CairoU256) -> Self {
+        Self { x, y }
+    }
+    pub fn from_bytes_be(x: &[u8; 32], y: &[u8; 32]) -> Self {
+        Self::new(
+            CairoU256::from_byte_slice_be(x),
+            CairoU256::from_byte_slice_be(y),
+        )
+    }
+}
+
+impl FeltSerialize for P256r1PublicKey {
+    fn to_felts(self) -> Vec<Felt252> {
+        self.x
+            .to_felts()
+            .into_iter()
+            .chain(self.y.to_felts())
+            .collect()
+    }
 }
 
 impl ArgsBuilder {
